JWT integrated

+++ routes/dealRoutes.js
@@ -1,7 +1,8 @@
-const mongoose = require('mongoose')
-const Misc = require('../utils/misc')
-const cb = require('../utils/callbacks')
-const constants = require('../config/constants')
+const JWT = require('../utils/jwt');
+const mongoose = require('mongoose');
+const Misc = require('../utils/misc');
+const cb = require('../utils/callbacks');
+const constants = require('../config/constants');
 
 var Deal = null;
 var User = null;
@@ -13,6 +14,10 @@ module.exports = function(app, dealsDB, usersDB) {
 
   // Create
   app.post('/deals', function(req, res) {
+    if (!JWT.verify(req.get("Bearer"))) {
+      return;
+    }
+
     const jsonData = req.body;
     const newID = mongoose.Types.ObjectId();
 
@@ -37,6 +42,10 @@ module.exports = function(app, dealsDB, usersDB) {
 
   // Read
   app.get('/deals', function(req, res) {
+    if (!JWT.verify(req.get("Bearer"))) {
+      return;
+    }
+
     const jsonData = req.body;
 
     if (Misc.isEmptyObject(jsonData)) {
@@ -49,6 +58,10 @@ module.exports = function(app, dealsDB, usersDB) {
 
   // Update
   app.put('/deals', function(req, res) {
+    if (!JWT.verify(req.get("Bearer"))) {
+      return;
+    }
+
     const jsonData = req.body;
     var id = jsonData.id;
     delete jsonData.id;
@@ -59,6 +72,10 @@ module.exports = function(app, dealsDB, usersDB) {
 
   // delete
   app.delete('/deals', function(req, res) {
+    if (!JWT.verify(req.get("Bearer"))) {
+      return;
+    }
+
     const jsonData = req.body;
 
     Deal.findByIdAndDelete(jsonData.id, (err, result) => cb.callback(res, err, result));
@@ -66,6 +83,10 @@ module.exports = function(app, dealsDB, usersDB) {
 
   // Increment the number of claims a deal has given a deal and user ID
   app.put('/deals/claim', function(req, res) {
+    if (!JWT.verify(req.get("Bearer"))) {
+      return;
+    }
+    
     const queryArgs = req.query;
     var dealID = null;
     var userID = null;
