+++ routes/userRoutes.js
@@ -1,6 +1,7 @@
-const mongoose = require('mongoose')
-const Misc = require('../utils/misc')
-const cb = require('../utils/callbacks')
+const JWT = require('../utils/jwt');
+const mongoose = require('mongoose');
+const Misc = require('../utils/misc');
+const cb = require('../utils/callbacks');
 
 var User = null;
 
@@ -11,6 +12,10 @@ module.exports = function(app, usersDB, dealsDB) {
 
   // Create
   app.post('/users/email', function(req, res) {
+    if (!JWT.verify(req.get("Bearer"))) {
+      return;
+    }
+
     const jsonData = req.body;
     const newID = mongoose.Types.ObjectId();
 
@@ -32,6 +37,10 @@ module.exports = function(app, usersDB, dealsDB) {
   });
 
   app.post('users/facebook', function(req, res) {
+    if (!JWT.verify(req.get("Bearer"))) {
+      return;
+    }
+
     const jsonData = req.body;
 
     var newObj = new User(
@@ -51,6 +60,10 @@ module.exports = function(app, usersDB, dealsDB) {
   });
 
   app.post('users/google', function(req, res) {
+    if (!JWT.verify(req.get("Bearer"))) {
+      return;
+    }
+
     const jsonData = req.body;
 
     var newObj = new User(
@@ -71,6 +84,10 @@ module.exports = function(app, usersDB, dealsDB) {
 
   // Read
   app.get('/users', function(req, res) {
+    if (!JWT.verify(req.get("Bearer"))) {
+      return;
+    }
+
     const jsonData = req.body;
 
     if (Misc.isEmptyObject(jsonData)) {
@@ -82,6 +99,10 @@ module.exports = function(app, usersDB, dealsDB) {
 
   // Update
   app.put('/users', function(req, res) {
+    if (!JWT.verify(req.get("Bearer"))) {
+      return;
+    }
+
     const jsonData = req.body;
     var id = jsonData.id;
     delete jsonData.id;
@@ -91,6 +112,10 @@ module.exports = function(app, usersDB, dealsDB) {
 
   // delete
   app.delete('/users', function(req, res) {
+    if (!JWT.verify(req.get("Bearer"))) {
+      return;
+    }
+
     const jsonData = req.body;
 
     User.findByIdAndDelete(jsonData.id, (err, result) => cb.callback(res, err, result));
@@ -99,6 +124,10 @@ module.exports = function(app, usersDB, dealsDB) {
   /****************************************************************************/
   // Get user profile by ID
   app.get('/user/profile', function(req, res) {
+    if (!JWT.verify(req.get("Bearer"))) {
+      return;
+    }
+
     const jsonData = req.body;
 
     User.findById(jsonData.id, (err, result) => cb.callback(res, err, result));
@@ -106,6 +135,10 @@ module.exports = function(app, usersDB, dealsDB) {
 
   // Get deals for a specific User
   app.get('/user/deals', function(req, res) {
+    if (!JWT.verify(req.get("Bearer"))) {
+      return;
+    }
+    
     const jsonData = req.body;
     const query = Misc.dealsQuery(jsonData);
